Improve configure for maemo plugin
ALSA bug#3860:
	https://bugtrack.alsa-project.org/alsa-bug/view.php?id=3860

The Maemo DSP plugin checks for D-Bus in configure.in and then makes a bold assumption that this means it should use a proprietary resource manager available only on a specific proprietary platform.

Attaching a patch to add --enable-maemo-resource-manager configure flag that enables the resource manager if set and if D-Bus is present.

Signed-off-by: Takashi Iwai <tiwai@suse.de>