jack: Restore PID to jack client name when not specified explicitly
Commit 14190cd8 added a facility to explicitly name Jack clients with
a 'name' option, but did so in a way that was not fully
backwards-compatible with older versions of the alsa-plugins. There
were two issues:

1. Re-using the magical "name" variable, which comes preloaded with
the name of the PCM

2. The code assumed snd_config_get_string sets a variable to NULL if
the config entry does not exist. Instead, this function leaves the
variable unchanged.

The result of these two errors meant that the PCM name would be
treated as if it were a user-defined name, and the old code which
embedded the PID would never be run

This code uses a completely new variable to store the explicitly-set
client name, so that the two code paths no longer collide.

[fixed trivial coding style issues by tiwai]

Signed-off-by: Branan Purvine-Riley <branan@gmail.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>