pph: Fix build with older speex
The previous commit fixed the build of the recent speexdsp, but it
also broke the build with the old library.  Actually the only needed
change is the inclusion of speex/speex_types.h instead of
speex/speexdsp_types.h, so it can be easily worked around via the
check in configure script.

Along with this change, make the version check back to 1.2 from
1.2rc2, as 1.2rc1 works fine now again, too.

Signed-off-by: Takashi Iwai <tiwai@suse.de>