upmix: generalize format, code cleanup
The main purpose of this change is to be able to change the format to
e.g. S32 instead of S16 by changing only two lines on the top of the source.
So a new symbol UPMIX_PCM_FORMAT is added to hide the actual format,
and a special type upmix_sample_t marks the sample type, instead of
fixed short.

The average_copy function was expecting the nchns parameter to be 2,
this parameter is now a local constant.

The variable i is renamed to channel as it denotes the channel index
in every case.

The behavior of the plugin remains identical.

Signed-off-by: Benedek Kupper <benedek.kupper@streamunlimited.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>