pcm_a52: Don't move bytes within the outbuf
The output buffer will never be appended, but a new a52 frame will always
be written in its entirety to outbuf. Therefore we don't need to move
bytes that were not yet sent to the slave to the beginning of the output
buffer.

Also don't overwrite the output buffer when there are still frames
to be sent.

BugLink: https://github.com/alsa-project/alsa-plugins/pull/8
Signed-off-by: Alexander Motzkau <a.motzkau@web.de>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>