jack: implement pcm operation sw_params
Alsa-lib commit 88e4ae27 introduced a dynamic recalculation of the avail_min
value on the underlaying slave devices.
This updated avail_min value is not reaching the alsa-jack plugin due to lack
of the sw_params operation.
This patch adds this missing operation to correctly apply an updated avail_min
setting during runtime.

Signed-off-by: Andreas Pape <apape@de.adit-jv.com>