samplerate: Make LGPL-only
As libsamplerate itself was already relicensed to 2-Clause BSD
license, there is no need for restriction of GPL license for
samplerate plugin, too.

Let's switch to LGPL-only for relaxing the use.

Luckily, I am the sole author of this piece of code, so it should be
no legal problem for immediate relicensing.

Closes: https://github.com/alsa-project/alsa-plugins/issues/54
Signed-off-by: Takashi Iwai <tiwai@suse.de>